# server_source/reports/openapi_manual_export.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Murim Inn Rebuild Server",
    "version": "0.4.0"
  },
  "paths": {
...
    "/monsters": {
      "get": {
        "tags": [
          "Monster"
        ],
        "summary": "List Monsters",
        "operationId": "list_monsters_monsters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Monster"
        ],
        "summary": "Upsert Monster",
        "operationId": "upsert_monster_monsters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Monster"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/monsters/{monster_id}": {
      "get": {
        "tags": [
          "Monster"
        ],
        "summary": "Get Monster",
        "operationId": "get_monster_monsters__monster_id__get",
        "parameters": [
          {
            "name": "monster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Monster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Monster"
        ],
        "summary": "Delete Monster",
        "operationId": "delete_monster_monsters__monster_id__delete",
        "parameters": [
          {
            "name": "monster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Monster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/iap/google/verify": {
      "post": {
        "tags": [
          "IAP",
          "Legacy"
        ],
        "summary": "Verify Google",
        "operationId": "verify_google_iap_google_verify_post",
        "parameters": [
          {
            "name": "X-Req-TS",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Req-Ts"
            }
          },
          {
            "name": "X-Req-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Req-Nonce"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GoogleVerifyReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/iap/apple/verify": {
      "post": {
        "tags": [
          "IAP",
          "Legacy"
        ],
        "summary": "Verify Apple",
        "operationId": "verify_apple_iap_apple_verify_post",
        "parameters": [
          {
            "name": "X-Req-TS",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Req-Ts"
            }
          },
          {
            "name": "X-Req-Nonce",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Req-Nonce"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AppleVerifyReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AppleVerifyReq": {
        "properties": {
          "productId": {
            "type": "string",
            "title": "Productid"
          },
          "txId": {
            "type": "string",
            "title": "Txid"
          },
          "raw": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "productId",
          "txId"
        ],
        "title": "AppleVerifyReq"
      },
...
      "GoogleVerifyReq": {
        "properties": {
          "productId": {
            "type": "string",
            "title": "Productid"
          },
          "purchaseToken": {
            "type": "string",
            "title": "Purchasetoken"
          },
          "txId": {
            "type": "string",
            "title": "Txid"
          },
          "raw": {
            "additionalProperties": true,
            "type": "object",
            "title": "Raw"
          }
        },
        "type": "object",
        "required": [
          "productId",
          "purchaseToken",
          "txId"
        ],
        "title": "GoogleVerifyReq"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Monster": {
        "properties": {
          "monsterId": {
            "type": "string",
            "title": "Monsterid"
          },
          "name": {
            "type": "string",
            "title": "Name",
            "default": ""
          },
          "level": {
            "type": "integer",
            "title": "Level",
            "default": 1
          },
          "hp": {
            "type": "number",
            "title": "Hp",
            "default": 0
          },
          "atk": {
            "type": "number",
            "title": "Atk",
            "default": 0
          },
          "defense": {
            "type": "number",
            "title": "Defense",
            "default": 0
          },
          "goldPerMin": {
            "type": "integer",
            "title": "Goldpermin",
            "default": 0
          },
          "expPerMin": {
            "type": "integer",
            "title": "Exppermin",
            "default": 0
          },
          "tags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Tags"
          }
        },
        "type": "object",
        "required": [
          "monsterId"
        ],
        "title": "Monster"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}